# [improvement][performance] improve lru cache resize performance and memory usage