# Proposed changes

## Problem Summary:

Previously we use doubly linked list for LRUCache's HandleTable, in order to support a lookup-free `remove(const LRUHandle* h)` function. However it 1) slows down the resize operation because we need to malloc/free the dummy LRUHandle node for each bucket 2) increases the overall memory usage.

It turns out that we can achieve fast `remove(const LRUHandle* h)` using just singly linked list.

## Checklist(Required)

1. Does it affect the original behavior: (No)
2. Has unit tests been added: (Yes)
3. Has document been added or modified: (No Need)
4. Does it need to update dependencies: (No)
5. Are there any changes that cannot be rolled back: (No)

## Further comments

If this is a relatively large or complex change, kick off the discussion at [dev@doris.apache.org](mailto:dev@doris.apache.org) by explaining why you chose the solution you did and what alternatives you considered, etc...