@xinyiZzz 2022-05-12
Hi @gaodayue 
Please provide test results to demonstrate the performance and memory usage benefits of singly linked list, especially at high concurrency.

The lru cache is used globally in Doris, and the resize is very low frequency. It will only be resized frequently when the BE is cold started.