@xinyiZzz 2022-05-17
> A bit of context for this PR: we observed several latency spikes after restarting BE and found that it's caused by the resize operation (see the graph below), so we'd like to improve the resize operation in order to avoid any query timeout.
> 
> The cluster uses 200G page cache per BE and resizing single shard to 8 million entries took >2 seconds.
> 
> I wrote a simple program to test the resize speed. From the results, we can see that for large resize, this PR reduces the resize duration by 60%. This PR also reduces the memory usage, but it's not our main concern.

Convincing test results, nice job~