@gaodayue 2022-05-19
Error from the Code Quality Analysis check, not related to this PR

```
[INFO] Doris FE Project Parent POM ........................ FAILURE [11:44 min]
[INFO] fe-common .......................................... SUCCESS [ 21.655 s]
[INFO] spark-dpp .......................................... SUCCESS [ 11.115 s]
[INFO] fe-core ............................................ SUCCESS [01:07 min]
[INFO] hive-udf ........................................... SUCCESS [ 19.401 s]
[INFO] java-udf ........................................... SUCCESS [ 46.357 s]
[INFO] ------------------------------------------------------------------------
[INFO] BUILD FAILURE
[INFO] ------------------------------------------------------------------------
[INFO] Total time:  14:33 min
[INFO] Finished at: 2022-05-18T04:06:40Z
[INFO] ------------------------------------------------------------------------
Error:  Failed to execute goal org.sonarsource.scanner.maven:sonar-maven-plugin:3.9.1.2184:sonar (default-cli) on project fe: Project not found. Please check the ‘sonar.projectKey’ and ‘sonar.organization’ properties, the ‘SONAR_TOKEN’ environment variable, or contact the project administrator -> [Help 1]
```